Move dtdl package to root level, delete test code
