+++ src/main/java/dqm/jku/dqmeerkat/dtdl/dto/ProfileStatisticDto.java
@@ -1,4 +1,4 @@
-package dqm.jku.dqmeerkat.domain.dtdl.dto;
+package dqm.jku.dqmeerkat.dtdl.dto;
 
 import lombok.Data;
 import lombok.experimental.SuperBuilder;
