+++ src/main/java/dqm/jku/dqmeerkat/dtdl/dto/RelationshipDto.java
@@ -1,7 +1,6 @@
-package dqm.jku.dqmeerkat.domain.dtdl.dto;
+package dqm.jku.dqmeerkat.dtdl.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.squareup.moshi.Json;
 import lombok.Builder;
 import lombok.Data;
