+++ src/main/resource/influx.properties
@@ -1,2 +1,3 @@
+# TODO gitignore me!
 db.token=12bdc4164c2e8141
 db.orgId=e43ef650859b33f0
